requirements.txt
# Core Django
Django==4.2.6
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7
django-db-geventpool==4.0.0  # For connection pooling in production

# Production Server
gunicorn==21.2.0
whitenoise==6.5.0  # For static files
uvicorn==0.23.2  # ASGI server (optional for async)

# Forms & UI
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-widget-tweaks==1.5.0  # For form field customization

# Email
django-anymail==10.1  # For production email (SendGrid, Mailgun, etc.)
mailhog==1.0.1  # For local email testing

# Security
django-axes==6.0.0  # For login security
django-csp==3.7  # Content Security Policy
django-ratelimit==3.0.1  # Rate limiting

# Async & Background Tasks
celery==5.3.4
redis==4.6.0  # Celery broker
flower==2.0.1  # Celery monitoring

# Monitoring
django-prometheus==2.3.1
sentry-sdk==1.34.0  # Error tracking

# API (if needed)
djangorestframework==3.14.0
drf-yasg==1.21.7  # Swagger/OpenAPI docs

# File Processing
Pillow==10.1.0  # Image handling
python-magic==0.4.27  # File type validation

# Date/Time
pytz==2023.3
python-dateutil==2.8.2

# Development Tools (not for production)
black==23.11.0
flake8==6.1.0
isort==5.12.0
django-debug-toolbar==4.1.0
pytest-django==4.5.2
factory-boy==3.2.1

# Production Dependencies
python-dotenv==1.0.0
brotlipy==0.7.0  # Brotli compression